pyproject.toml
[project]
name = "llm-suite"
version = "0.1.0"
description = "Universal tools for prompt engineering and LLM interactions."
authors = [
    {name = "Marek Piotr Mysior"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "anthropic>=0.50.0",
    "instructor>=1.7.9",
    "jinja2>=3.1.6",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-frontmatter>=1.1.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "langfuse>=2.60.3",
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llmsuite"]


[project.urls]
Homepage = "https://github.com/mmysior/llm-suite"
Issues = "https://github.com/mmysior/llm-suite/issues"